PLIM-42 Make Modify Password and Change Id type by Allergy

+++ src/main/java/com/plim/plimserver/domain/allergy/repository/CrossAllergyRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.plim.plimserver.domain.allergy.domain.CrossAllergy;
 
-public interface CrossAllergyRepository extends JpaRepository<CrossAllergy, Long>{
+public interface CrossAllergyRepository extends JpaRepository<CrossAllergy, Integer>{
 
 }
